fix: Address review feedback
